// src/api/generated/model/vehicleDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * GearFlow.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { VehicleDtoYearFabrication } from './vehicleDtoYearFabrication';
import type { VehicleDtoYearModel } from './vehicleDtoYearModel';

export interface VehicleDto {
  id: string;
  licensePlate: string;
  mark: string;
  model: string;
  color: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  yearFabrication: VehicleDtoYearFabrication;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  yearModel: VehicleDtoYearModel;
}